impletmented food api
Made a new page for food entry. it will confim the amount and name of the food it got from nutrition api before pushing to DB. and changing the name will reset the confirmation prompt. So the food nutr data will only be pushed to DB IF user confirms it.