Random qoute
I made it, so now a random (with seed) quote of the day will be generated and shown on the home screen. I will be using seed made by the date and the start and end of a user's name. Meaning it will be different from user to user, but remain the same throughout a day until the date is changed.